fix sample sliders when file size exceeds int32
now slider ticks are hardcoded to a max of 1000000000 if the filesize is
larger than that. the samples used will internally scale to the total
number of samples correctly.

also reworked the settings dict and the stringified one for the config
(which just calls str() on each value of the already parsed dict)